New: Customizeable action component styles 🎊
