+++ package.json
@@ -71,6 +71,7 @@
     "@storybook/theming": "5.3.18",
     "@testing-library/jest-dom": "5.5.0",
     "@testing-library/react": "10.0.3",
+    "@types/classnames": "2.2.10",
     "babel-loader": "8.1.0",
     "css-loader": "3.5.2",
     "eslint-config-eloquence": "13.7.0",
