+++ src/Button/Button.js
@@ -1,12 +1,18 @@
 import elem from '../elem-factory'
-import { actionClasses } from '../utils/componentry'
+import { actionClasses } from '../utils/classes'
 import { useTheme } from '../Theme/Theme'
 
 /**
  * [Button component 📝](https://componentry.design/components/button)
  */
 export default function Button(props) {
-  const { as = 'button', type = 'button', variant = 'primary', ...merged } = {
+  const {
+    as = 'button',
+    baseClass = 'btn',
+    type = 'button',
+    variant = 'primary',
+    ...merged
+  } = {
     ...useTheme('Button'),
     ...props,
   }
@@ -14,8 +20,10 @@ export default function Button(props) {
   return elem({
     as,
     type,
-    elemClassName: actionClasses('btn', variant, merged),
+    elemClassName: actionClasses(baseClass, variant, merged),
     ...merged,
   })
 }
-Button.displayName = 'Button'
+Button.displayName = '✨Button'
+/** Base class for creating Button className variants */
+Button.baseClass = 'btn'
