+++ src/Intro.stories.mdx
@@ -10,6 +10,12 @@ import { Button, Heading, Theme } from './index'
 
 _This story has examples of the core library component APIs and conventions._
 
+### Terms
+
+- Action: an operating mechanism, or the manner in which a mechanism or instrument
+  operates. Use with components that have elements which cause a change of state when
+  triggered.
+
 ### Render as
 
 _All components accept an `as` prop that will override the container element passed to
