+++ src/Nav/Nav.js
@@ -1,6 +1,7 @@
 import elem from '../elem-factory'
 import { useTheme } from '../Theme/Theme'
-import { actionClasses, navClasses } from '../utils/componentry'
+import { navClasses } from '../utils/componentry'
+import { actionClasses } from '../utils/classes'
 
 /**
  * [Nav component 📝](https://componentry.design/components/nav)
@@ -20,14 +21,14 @@ Nav.displayName = 'Nav'
  * [Nav item component 📝](https://componentry.design/components/nav)
  */
 Nav.Item = function NavItem(props) {
-  const { variant = 'nav-item', ...merged } = {
+  const { baseClass = 'nav-item', variant = 'nav-item', ...merged } = {
     ...useTheme('NavItem'),
     ...props,
   }
 
   return elem({
     as: 'a',
-    elemClassName: actionClasses(variant, merged),
+    elemClassName: actionClasses(baseClass, variant, merged),
     ...merged,
   })
 }
